refactor: Move template prefill loading into CreateAgentForm
