+++ frontend/apps/web/src/routes/CreateAgentPage.tsx
@@ -1,31 +1,11 @@
-import { useProjectMachinePoolGrants, useToolCatalog } from '@omnara/react'
-import { useSearch } from '@tanstack/react-router'
-
-import { agentTemplates } from '@/components/agents/agentTemplates'
 import { CreateAgentForm } from '@/components/agents/CreateAgentForm'
-import { Button } from '@/components/ui/button'
-import { FullPageSpinner, Spinner } from '@/components/ui/spinner'
-import { errorMessage } from '@/lib/submit-status'
+import { FullPageSpinner } from '@/components/ui/spinner'
 import { useProjectPage } from '@/lib/use-project-page'
 
 export function CreateAgentPage() {
-  const { activeOrg, project, projectId, isPending: projectIsPending } = useProjectPage()
-  const search = useSearch({ strict: false })
+  const { project, isPending } = useProjectPage()
 
-  const toolCatalog = useToolCatalog()
-  const canManage = project?.access.can_manage === true
-  const poolGrantsQuery = useProjectMachinePoolGrants(activeOrg.id, projectId, {
-    sort: 'name',
-    pageSize: 1,
-    enabled: canManage,
-  })
-  const catalog = toolCatalog.data
-  const defaultPool = poolGrantsQuery.data?.pages[0]?.data[0]?.machine_pool
-  const templatesError = toolCatalog.isError || poolGrantsQuery.isError
-  const templatesReady = !templatesError && !toolCatalog.isPending && !poolGrantsQuery.isPending
-  const linkedTemplate = agentTemplates.find((template) => template.id === search.template)
-
-  if (projectIsPending) return <FullPageSpinner />
+  if (isPending) return <FullPageSpinner />
 
   if (!project?.access.can_manage) {
     return (
@@ -42,51 +22,5 @@ export function CreateAgentPage() {
     )
   }
 
-  if (linkedTemplate && templatesError) {
-    const retrying = toolCatalog.isFetching || poolGrantsQuery.isFetching
-    return (
-      <div className="mx-auto flex w-full max-w-5xl flex-col gap-2">
-        <div
-          role="alert"
-          className="border-destructive/30 bg-destructive/5 flex items-center justify-between gap-3 rounded-md border px-3 py-2"
-        >
-          <p className="text-destructive text-sm">
-            {errorMessage(
-              toolCatalog.error ?? poolGrantsQuery.error,
-              'Could not load the template defaults.',
-            )}
-          </p>
-          <Button
-            type="button"
-            size="sm"
-            variant="outline"
-            className="shrink-0"
-            disabled={retrying}
-            onClick={() => {
-              if (toolCatalog.isError) void toolCatalog.refetch()
-              if (poolGrantsQuery.isError) void poolGrantsQuery.refetch()
-            }}
-          >
-            {retrying && <Spinner />}
-            Retry
-          </Button>
-        </div>
-      </div>
-    )
-  }
-
-  if (linkedTemplate && !templatesReady) return <FullPageSpinner />
-
-  return (
-    <CreateAgentForm
-      orgId={activeOrg.id}
-      orgName={activeOrg.name}
-      projectId={projectId}
-      projectName={project.name}
-      catalog={catalog}
-      defaultPool={defaultPool}
-      templatesReady={templatesReady}
-      initialTemplate={linkedTemplate}
-    />
-  )
+  return <CreateAgentForm />
 }
